Move free-standing fns into utils
